Rebuild if llama.cpp source changes / minor build speedup
Instead of cp -r / robocopy, build from the source directory. This
mildly speeds up the build although probably not noticeable on NVME
drives. The cargo-cmake crate will automatically place output in the
out/ folder for us.

Additionally, walk the source tree to tell cargo that a rebuild is
necessary if anything changes from the source. This ensures that changes
in the llama.cpp code trigger a rebuild which makes hacking on things a
bit easier.

Looks like this copying logic was copied from sherpa-onnx given the
comments seem to be copy-pasted so remove those references.