Pure refactor - cleanup build script
The OS detection code was bothering me as it wasn't properly doing cross
compilation (some places were and some weren't). Additionally, the OS
detection was a bit haphazard. This is a pure cleanup that parses the
information in TARGET up-front into an enum that is then checked instead
of working with strings.